pb: put uploadCoursesCsv in its own file, all tests now pass
