+++ javascript/src/main/services/Courses/CourseService.js
@@ -54,21 +54,4 @@ const buildDeleteCourse = (getToken, onSuccess, onError) => {
   return func;
 }
 
-const uploadCoursesCSV = (getToken, onSuccess, onError) =>{
-  const func = async (file) => {
-      const data = new FormData();
-      data.append("csv", file);
-      try{
-          await fetchWithToken(`/api/courses/upload`, getToken, {
-              method: "POST",
-              body: data
-          });
-          onSuccess();
-      } catch (err){
-          onError(err);
-      }
-  };
-  return func;
-}
-
-export { uploadCoursesCSV, buildCreateCourse, buildDeleteCourse, buildUpdateCourse };
+export { buildCreateCourse, buildDeleteCourse, buildUpdateCourse };
